+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Fixes:
 - Fixed dead entities sometimes not being counted as dead (hopefully)
+- Fixed double hook messages not having color (when not in a party)
 
 ### Back-end:
 - Added debug entities command